adjust price variable in products table

+++ db/schema/02_products.sql
@@ -7,6 +7,6 @@ CREATE TABLE products (
   description TEXT NOT NULL,
   photo_url VARCHAR(255) NOT NULL,
   calories_per_serving INTEGER NOT NULL,
-  price INTEGER NOT NULL,
+  price DECIMAL(10,2) NOT NULL,
   category VARCHAR(255) NOT NULL
 );
